make provider as module
