+++ flask_project/campaign_manager/campaigns_data/campaign/96224fd319834166b3dbc564d9e98a13.json
@@ -1,84 +1,1 @@
-{
-  "campaign_creator": "meomancer ",
-  "campaign_managers": [
-    "meomancer"
-  ],
-  "campaign_status": "Start",
-  "coverage": "",
-  "description": "",
-  "edited_at": "Mon May 22 13:59:28 2017",
-  "edited_by": "meomancer",
-  "end_date": null,
-  "geometry": {
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "type": "Polygon",
-          "coordinates": [
-            [
-              [
-                20.430364608764652,
-                -34.02157746626568
-              ],
-              [
-                20.439076423645023,
-                -34.03064717674483
-              ],
-              [
-                20.45040607452393,
-                -34.02506319517012
-              ],
-              [
-                20.45040607452393,
-                -34.02349819173851
-              ],
-              [
-                20.434269905090336,
-                -34.0185184433697
-              ],
-              [
-                20.430364608764652,
-                -34.02157746626568
-              ]
-            ]
-          ]
-        }
-      }
-    ]
-  },
-  "name": "Campaign 1",
-  "selected_functions": {
-    "function-1": {
-      "function": "CountFeature",
-      "feature": "buildings",
-      "category": "quality-function",
-      "attributes": ""
-    },
-    "function-2": {
-      "function": "FeatureAttributeCompleteness",
-      "feature": "buildings",
-      "category": "quality-function",
-      "attributes": "name=;"
-    },
-    "function-4": {
-      "function": "UploadShapefile",
-      "feature": null,
-      "category": "coverage-function",
-      "attributes": ""
-    },
-    "function-3": {
-      "function": "FeatureAttributeCompletenessError",
-      "feature": "buildings",
-      "category": "error-function",
-      "attributes": "name=;building=;"
-    }
-  },
-  "start_date": "2017-05-01",
-  "tags": [],
-  "uploader": "meomancer",
-  "uuid": "96224fd319834166b3dbc564d9e98a13",
-  "version": 25
-}
+{"campaign_creator": "meomancer ", "campaign_managers": ["meomancer"], "campaign_status": "Start", "coverage": "", "description": "", "edited_at": "Mon May 22 13:59:28 2017", "edited_by": "meomancer", "end_date": null, "geometry": {"type": "FeatureCollection", "features": [{"type": "Feature", "properties": {}, "geometry": {"type": "Polygon", "coordinates": [[[20.430364608764652, -34.02157746626568], [20.439076423645023, -34.03064717674483], [20.45040607452393, -34.02506319517012], [20.45040607452393, -34.02349819173851], [20.434269905090336, -34.0185184433697], [20.430364608764652, -34.02157746626568]]]}}]}, "name": "Campaign 1", "selected_functions": {"function-1": {"function": "CountFeature", "feature": "buildings", "category": "quality-function", "attributes": ""}, "function-2": {"function": "FeatureAttributeCompleteness", "feature": "buildings", "category": "quality-function", "attributes": "name=;"}, "function-3": {"function": "UploadCoverage", "feature": null, "category": "coverage-function", "attributes": ""}}, "start_date": "2017-05-01", "tags": [], "uploader": "meomancer", "uuid": "96224fd319834166b3dbc564d9e98a13", "version": 26}
