+++ flask_project/campaign_manager/campaigns_data/campaign/96224fd319834166b3dbc564d9e98a15.json
@@ -1,1 +1,1 @@
-{"campaign_creator": "dimasciput ", "campaign_managers": ["dimasciput", "meomancer"], "campaign_status": "Start", "coverage": "", "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut efficitur eleifend nulla, et tincidunt urna accumsan vitae. Proin posuere sed dui quis lobortis.", "edited_at": "Mon May 22 13:59:28 2017", "edited_by": "meomancer", "end_date": "2017-09-15", "geometry": {"type": "FeatureCollection", "features": [{"type": "Feature", "properties": {}, "geometry": {"type": "Polygon", "coordinates": [[[112.50635147094728, -7.0862427881576115], [112.48609542846681, -7.201725645957958], [112.5578498840332, -7.219777841831864], [112.57020950317384, -7.233061073084732], [112.61003494262697, -7.221821441281233], [112.62754440307619, -7.212284564870234], [112.64265060424805, -7.149608677183668], [112.64265060424805, -7.108047091883416], [112.60866165161134, -7.079088026071719], [112.54755020141603, -7.072955284539926], [112.51630783081055, -7.07568095750246], [112.50635147094728, -7.0862427881576115]]]}}]}, "name": "Campaign 2", "selected_functions": {"function-1": {"function": "CountFeature", "feature": "buildings", "category": "quality-function", "attributes": ""}, "function-2": {"function": "FeatureAttributeCompleteness", "feature": "buildings", "category": "quality-function", "attributes": "name=;amenity=hospital,pharmacy"}}, "start_date": "2017-05-01", "tags": ["boundaries", "buildings", "educational facilities"], "uploader": "meomancer", "uuid": "96224fd319834166b3dbc564d9e98a15", "version": 26}
+{"campaign_creator": "dimasciput ", "campaign_managers": ["dimasciput", "meomancer"], "campaign_status": "Start", "coverage": "", "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut efficitur eleifend nulla, et tincidunt urna accumsan vitae. Proin posuere sed dui quis lobortis.", "edited_at": "Mon May 22 13:59:28 2017", "edited_by": "meomancer", "end_date": "2017-09-15", "geometry": {"type": "FeatureCollection", "features": [{"type": "Feature", "properties": {}, "geometry": {"type": "Polygon", "coordinates": [[[112.50635147094728, -7.0862427881576115], [112.48609542846681, -7.201725645957958], [112.5578498840332, -7.219777841831864], [112.57020950317384, -7.233061073084732], [112.61003494262697, -7.221821441281233], [112.62754440307619, -7.212284564870234], [112.64265060424805, -7.149608677183668], [112.64265060424805, -7.108047091883416], [112.60866165161134, -7.079088026071719], [112.54755020141603, -7.072955284539926], [112.51630783081055, -7.07568095750246], [112.50635147094728, -7.0862427881576115]]]}}]}, "name": "Campaign 2", "selected_functions": {"function-1": {"function": "FeatureAttributeCompleteness", "feature": "buildings", "category": "quality-function", "attributes": "name="}, "function-2": {"function": "CountFeature", "feature": "buildings", "category": "quality-function", "attributes": ""}}, "start_date": "2017-05-01", "tags": ["boundaries", "buildings", "educational facilities"], "uploader": "meomancer", "uuid": "96224fd319834166b3dbc564d9e98a15", "version": 28}
