+++ flask_project/campaign_manager/data_providers/__init__.py
@@ -1,0 +1,2 @@
+__author__ = 'Irwan Fathurrahman <irwan@kartoza.com>'
+__date__ = '29/05/17'
