+++ flask_project/campaign_manager/insights_functions/__init__.py
@@ -3,4 +3,4 @@ __date__ = '17/05/17'
 
 from .count_feature import CountFeature
 from .feature_attribute_completeness import FeatureAttributeCompleteness
-from .upload_shapefile import UploadShapefile
+from .upload_coverage import UploadCoverage
