+++ flask_project/campaign_manager/insights_functions/count_feature.py
@@ -1,23 +1,21 @@
 __author__ = 'Irwan Fathurrahman <irwan@kartoza.com>'
 __date__ = '17/05/17'
-
-from campaign_manager.selected_functions._abstract_insights_function import (
-    AbstractInsightsFunction
+from campaign_manager.insights_functions._abstract_overpass_insight_function import (
+    AbstractOverpassInsightFunction
 )
 
 
-class FeatureAttributeCompleteness(AbstractInsightsFunction):
-    function_name = "Showing feature completeness"
+class CountFeature(AbstractOverpassInsightFunction):
+    function_name = "Showing number of feature in group"
     category = ['quality']
-    need_required_attributes = True
-    icon = 'list'
+    need_required_attributes = False
 
     def get_ui_html_file(self):
         """ Get ui name in templates
         :return: string name of html
         :rtype: str
         """
-        return "feature_completeness"
+        return "piechart"
 
     def get_summary_html_file(self):
         """ Get summary name in templates
@@ -43,16 +41,19 @@ class FeatureAttributeCompleteness(AbstractInsightsFunction):
         :return: Processed data
         :rtype: dict
         """
-        metadata = self.metadata()
-        required_attributes = {}
-        required_attributes.update(self.get_required_attributes())
-        output = {
-            'attributes': required_attributes,
-            'data': self._function_good_data,
-            'percentage': '%.1f' % (
-                (len(data) / metadata['collected_data_count']) * 100
-            ),
-            'complete': len(data),
-            'total': metadata['collected_data_count']
-        }
+        output = {}
+        for current_data in data:
+            building_key = 'building'
+            if building_key not in current_data:
+                building_key = 'amenity'
+
+            try:
+                building_type = current_data[building_key]
+            except KeyError:
+                building_type = 'unknown'
+
+            if building_type not in output:
+                output[building_type] = 0
+            output[building_type] += 1
+
         return output
