--- flask_project/campaign_manager/selected_functions/upload_shapefile.py
@@ -1,55 +1,0 @@
-__author__ = 'Irwan Fathurrahman <irwan@kartoza.com>'
-__date__ = '17/05/17'
-
-from campaign_manager.selected_functions._abstract_insights_function import (
-    AbstractInsightsFunction
-)
-
-
-class UploadShapefile(AbstractInsightsFunction):
-    function_name = "Upload Shapefile for creating coverage"
-    category = ['coverage']
-    need_feature = False
-    need_required_attributes = False
-
-    def get_ui_html_file(self):
-        """ Get ui name in templates
-        :return: string name of html
-        :rtype: str
-        """
-        return "upload_coverage"
-
-    def get_summary_html_file(self):
-        """ Get summary name in templates
-        :return: string name of html
-        :rtype: str
-        """
-        return ""
-
-    def get_details_html_file(self):
-        """ Get summary name in templates
-        :return: string name of html
-        :rtype: str
-        """
-        return ""
-
-    def post_process_data(self, data):
-        """ Process data regarding output.
-        This needed for processing data for counting or grouping.
-
-        :param data: Data that received from open street map
-        :type data: dict
-
-        :return: Processed data
-        :rtype: dict
-        """
-        output = {
-            'uuid': self.campaign.uuid,
-            'files': self.campaign.get_coverage_files(),
-        }
-        if self.campaign.coverage:
-            output['coverage'] = {
-                'last_uploader': self.campaign.coverage['last_uploader'],
-                'last_uploaded': self.campaign.coverage['last_uploaded']
-            }
-        return output
