+++ flask_project/campaign_manager/templates/campaign_widget/ui/upload_coverage.html
@@ -59,7 +59,14 @@
 
 </style>
 <script>
+    var coverage_geojson = null;
+    {% if data.coverage %}
+        coverage_geojson = {{ data.coverage.geojson|safe }};
+    {% endif %}
     var is_in_upload = false;
+    var coverage_layer = null;
+    var coverage_info = null;
+    var coverage_legend = null;
     function added_file_indicator(filename, complete) {
         var complete_class = '';
         var percent = '0%';
@@ -83,7 +90,7 @@
             $('#coverage-last-upload').hide();
         });
         $(".file-upload").click(function () {
-            if ($(".coverage-uploader #upload-button").get(0).files.length > 0) {
+            if (is_in_upload && $('.file-indicator').find('.file-indicator-row ').length > 0) {
                 $(this).prop('disabled', true);
                 $(".coverage-uploader .add-file").prop('disabled', true);
                 $(".coverage-uploader #upload-button").css('cursor', 'not-allowed');
@@ -128,7 +135,6 @@
                     $(".coverage-uploader #upload-button").prop('disabled', false);
                     is_in_upload = false;
                     upload_complete();
-                    get_coverage();
                 }
             },
             progress: function (e, uploadData) {
@@ -138,34 +144,126 @@
                 $nameIndicator.find('.percent-indicator').html(progress + '%');
             }
         });
-        function upload_complete() {
-            $.ajax({
-                url: '/campaign_manager/campaign/' + uuid + '/coverage-upload-success?uploader=' + $("#profile-name").html(),
-                dataType: 'json',
-                success: function (data) {
-                    if (!data['success']) {
-                        $('#coverage-last-upload').show();
-                        $('#coverage-last-upload').css('color', 'red');
-                        $('#coverage-last-upload').html(data['reason']);
-                    } else {
-                        console.log(data);
-                        $(".file-indicator").html('');
-                        $('#coverage-last-upload').show();
-                        $('#coverage-last-upload').css('color', 'gray');
-                        $('#coverage-last-upload').html(
-                                'uploaded by <a href="www.openstreetmap.org/user/' +
-                                data['data']['last_uploader'] + '">' +
-                                data['data']['last_uploader'] + '</a> at ' +
-                                data['data']['last_uploaded']);
-                        $.each(data['files'], function (index, value) {
-                            added_file_indicator(value, true);
-                        });
+        get_coverage();
+    });
+    function upload_complete() {
+        $.ajax({
+            url: '/campaign_manager/campaign/' + uuid + '/coverage-upload-success?uploader=' + $("#profile-name").html(),
+            dataType: 'json',
+            success: function (data) {
+                if (!data['success']) {
+                    $('#coverage-last-upload').show();
+                    $('#coverage-last-upload').css('color', 'red');
+                    $('#coverage-last-upload').html(data['reason']);
+                } else {
+                    $(".file-indicator").html('');
+                    $('#coverage-last-upload').show();
+                    $('#coverage-last-upload').css('color', 'gray');
+                    $('#coverage-last-upload').html(
+                            'uploaded by <a href="www.openstreetmap.org/user/' +
+                            data['data']['last_uploader'] + '">' +
+                            data['data']['last_uploader'] + '</a> at ' +
+                            data['data']['last_uploaded']);
+                    $.each(data['files'], function (index, value) {
+                        added_file_indicator(value, true);
+                    });
+                    coverage_geojson = data['data']['geojson'];
+                    get_coverage();
+                }
+            }
+        });
+    }
+    function get_coverage() {
+        if (!coverage_geojson) {
+            return;
+        }
+        var that = this;
+        if (Object.keys(coverage_geojson).length != 0) {
+            if (coverage_layer) {
+                that.map.removeLayer(coverage_layer);
+                that.map.removeLayer(coverage_info);
+                that.map.removeLayer(coverage_legend);
+            }
 
+            var today = new Date();
+            var style = {
+                weight: 2,
+                opacity: 1,
+                color: 'white',
+                dashArray: '3',
+                fillOpacity: 0.8
+            };
+            var older_color = "#5CB85C";
+            var future_color = "#E76F4A";
+
+            {# create layer #}
+            coverage_layer = L.geoJSON(coverage_geojson, {
+                style: function (feature) {
+                    var property_date = new Date(feature.properties.date);
+                    if (property_date < today) {
+                        style['fillColor'] = older_color;
+                        return style;
+                    } else {
+                        style['fillColor'] = future_color;
+                        return style;
                     }
+                },
+                onEachFeature: function (feature, layer) {
+                    layer.on({
+                        mouseover: highlightFeature,
+                        mouseout: resetHighlight,
+                        click: zoomToFeature
+                    });
+                },
+                filter: function (feature) {
+                    if (!feature.properties.date) return false;
+                    return true;
                 }
             });
+            that.map.addLayer(coverage_layer);
+
+            {# create legend #}
+            coverage_legend = L.control(
+                    {position: 'bottomright'}
+            );
+            coverage_legend.onAdd = function () {
+                var div = L.DomUtil.create('div', 'info legend');
+                div.innerHTML += '<div><i style="background:' + older_color + '"></i>Mapped<br></div>';
+                div.innerHTML += '<div><i style="background:' + future_color + '"></i>Not yet<br></div>';
+                return div;
+            };
+            coverage_legend.addTo(that.map);
+
+            {# create info #}
+            coverage_info = L.control(
+                    {position: 'topright'}
+            );
+            coverage_info.onAdd = function () {
+                this._div = L.DomUtil.create('div', 'info');
+                $(this._div).hide();
+                this.update();
+                return this._div;
+            };
+            coverage_info.update = function (properties) {
+                $(this._div).hide();
+                if (properties) {
+                    $(this._div).show();
+                    var html = '<h4>Area detail</h4>';
+                    html += '<list-group>';
+                    $.each(properties, function (key, property) {
+                        html += '<div class="list-group-item">' + key.toLowerCase() +
+                                '<span class="pull-right text-muted">' + capitalizeFirstLetter(property.toLowerCase()) +
+                                '</span>' +
+                                '</div>';
+                    });
+                    html += '</list-group>';
+                    $(this._div).html(html);
+                }
+            };
+
+            coverage_info.addTo(that.map);
         }
-    });
+    }
 </script>
 <script>
     {# browser event #}
